chatbot/requirements.txt
streamlit_chat
langchain_core